ref : 통합테스트에 정의된 단위테스트성격 테스트메서드들 -> 단위테스트 파일로 이동 리팩토링
