+++ src/test/java/com/even/zaro/service/NotificationServiceTest.java
@@ -3,6 +3,8 @@ package com.even.zaro.service;
 import com.even.zaro.dto.notification.NotificationDto;
 import com.even.zaro.entity.Notification;
 import com.even.zaro.entity.User;
+import com.even.zaro.global.ErrorCode;
+import com.even.zaro.global.exception.notification.NotificationException;
 import com.even.zaro.global.util.NotificationMapper;
 import com.even.zaro.repository.NotificationRepository;
 import com.even.zaro.repository.UserRepository;
@@ -19,6 +21,7 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -108,6 +111,26 @@ public class NotificationServiceTest {
     class MarkAsReadTest {
 
         @Test
+        void 알림_1개_읽음_처리_성공() {
+            Notification noti = createNotification(user, Notification.Type.LIKE, false);
+            when(notificationRepository.findByIdAndUserId(noti.getId(), userId)).thenReturn(Optional.of(noti));
+
+            notificationService.markAsRead(noti.getId(), userId);
+
+            assertThat(noti.isRead()).isTrue();
+        }
+
+        @Test
+        void 존재하지_않는_알림_읽음_처리_실패() {
+            when(notificationRepository.findByIdAndUserId(9999L, userId)).thenReturn(Optional.empty());
+
+            NotificationException exception = assertThrows(NotificationException.class, () ->
+                    notificationService.markAsRead(9999L, userId));
+
+            assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.NOTIFICATION_NOT_FOUND);
+        }
+
+        @Test
         void 이미_읽은_알림을_다시_읽으면_예외_없이_읽음_유지() {
             Notification alreadyRead = createNotification(user, Notification.Type.FOLLOW, true);
             when(notificationRepository.findByIdAndUserId(alreadyRead.getId(), userId)).thenReturn(Optional.of(alreadyRead));
@@ -121,6 +144,19 @@ public class NotificationServiceTest {
 
     @Nested
     class MarkAllAsReadTest {
+
+        @Test
+        void 알림_전체_읽음_처리_성공() {
+            Notification n1 = createNotification(user, Notification.Type.FOLLOW, false);
+            Notification n2 = createNotification(user, Notification.Type.LIKE, false);
+
+            when(notificationRepository.findAllByUserIdAndIsReadFalse(userId)).thenReturn(List.of(n1, n2));
+
+            notificationService.markAllAsRead(userId);
+
+            assertThat(n1.isRead()).isTrue();
+            assertThat(n2.isRead()).isTrue();
+        }
 
         @Test
         void 모든_알림이_이미_읽음상태면_예외없이_전체_유지() {
